Fix incorrect prerequisites
'Nothing to be done' build failed because

* $(LOCAL_MODULE) should not be a prerequisite
* The created symlink doesn't exist in the host out tree

Besides, all prerequisites of $(SYMLINKS) are removed. This is because
'make' checks the timestamp of the file the link references rather than
the link itself. Therefore $(LOCAL_INSTALLED_MODULE) is a useless
prerequisite. On the other hand, $(LOCAL_PATH)/Android.mk should not
be a prerequisite of $(SYMLINKS) unless it is also a prerequisite of
$(LOCAL_INSTALLED_MODULE). Otherwise, if $(LOCAL_PATH)/Android.mk is
newer than $(LOCAL_INSTALLED_MODULE), $(SYMLINKS) will always be
re-made.

Correct the prerequisites of $(SUPERUSER_RC) as well.

Change-Id: I5746e983046a9cdc3ee2b8ecd3827f1f7e6b596e